% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar.gen.R
\name{add.metar.episodes}
\alias{add.metar.episodes}
\title{Divide METAR overpasses into episodes}
\usage{
add.metar.episodes(df.metar)
}
\arguments{
\item{df.metar}{Data frame.}
}
\value{
Data frame with episode numbers (multiple overpasses on the
    same date) in ascending order
}
\description{
Remember to run the data frame through \code{mutate(date =
convert.date(time))} first
}
\examples{

}
