% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c-base.tune.R
\name{C-BASE_correct}
\alias{C-BASE_correct}
\alias{correct.cbase.lm}
\alias{correct.cbase.ident}
\title{C-BASE local cloud base correction functions}
\usage{
correct.cbase.lm(df, correction)

correct.cbase.ident(df, correction)
}
\arguments{
\item{df}{Data.frame.  Contains local cloud bases to be corrected}

\item{correction}{List produced by the \ref{C-BASE_tune} functions}
}
\value{
Data frame with corrected local cloud bases
}
\description{
C-BASE local cloud base correction functions
}
\section{Functions}{
\itemize{
\item \code{correct.cbase.lm}: Correct using linear model or inherited
classes (including SVM)

\item \code{correct.cbase.ident}: "Identity" (i.e., no) correction
}}

