% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{C-BASE_factors}
\alias{C-BASE_factors}
\alias{factor.cbase.tuning.dist}
\alias{factor.cbase.tuning.mult}
\alias{factor.cbase.tuning.thick}
\title{C-BASE tuning factors}
\usage{
factor.cbase.tuning.dist(df)

factor.cbase.tuning.mult(df)

factor.cbase.tuning.thick(df)
}
\value{
With the exception of \code{factor.vfm}, these functions
    return factor version of their character or integer vector
    argument.
}
\description{
C-BASE tuning factors
}
\section{Functions}{
\itemize{
\item \code{factor.cbase.tuning.dist}: Convert distance to factor

\item \code{factor.cbase.tuning.mult}: Convert retrieval multiplicity to factor

\item \code{factor.cbase.tuning.thick}: Convert cloud thickness to factor
}}

