% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c-base.gen.R
\name{C-BASE_gen}
\alias{C-BASE_gen}
\alias{bases.cbase}
\alias{bases.cbase.dardar}
\alias{bases.2b.geoprof.lidar}
\title{C-BASE generator functions}
\usage{
bases.cbase(path = "/home/jmuelmen/CALIOP/VFM.v4.10/2008",
  pattern = "CAL_LID_L2_VFM-Standard-V4-10.*hdf",
  out.name = "cloud-bases.rds", ncores = 72)


  bases.cbase.dardar(path = "/projekt3/climate/DATA/SATELLITE/MULTI_SENSOR/DARDAR/DARDAR_MASK/2007",
  out.name = "cloud-bases.rds")


  bases.2b.geoprof.lidar(path = "/projekt3/climate/DATA/SATELLITE/CLOUDSAT/2B-GEOPROF-LIDAR/2008")
}
\arguments{
\item{path}{Character.  Apply algorithm to all HDF files in this path}

\item{out.name}{Character.  Output file name}
}
\value{
A data frame containing cloud bases
}
\description{
C-BASE generator functions
}
\section{Functions}{
\itemize{
\item \code{bases.cbase}: Bases by non-attenuated CALIOP in thin
clouds (using CALIOP VFM)

\item \code{bases.cbase.dardar}: Bases by non-attenuated CALIOP in thin clouds (using DARDAR mask)

\item \code{bases.2b.geoprof.lidar}: Bases by 2B-GEOPROF-LIDAR CloudSat/CALIOP combination
}}

