% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c-base.tune.R
\name{C-BASE_tune}
\alias{C-BASE_tune}
\alias{tune.cbase.lm}
\title{C-BASE tuning functions}
\usage{
tune.cbase.lm(df)
}
\arguments{
\item{df}{Data.frame.  Contains tuning data set}
}
\value{
A list with two components, a list of tuned models for each
    data category and a data.frame containing the index into said list of models
}
\description{
C-BASE tuning functions
}
\section{Functions}{
\itemize{
\item \code{tune.cbase.lm}: Tuning using linear models
}}

