% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{CALIOP_VFM_factors}
\alias{CALIOP_VFM_factors}
\alias{factor.feature.type}
\alias{factor.horizontal.averaging}
\alias{factor.ice.water.phase}
\alias{factor.qa}
\alias{factor.relabel}
\alias{factor.vfm}
\title{CALIOP VFM factors}
\usage{
factor.relabel(x, labels, ordered = FALSE)

factor.feature.type(x)

factor.qa(x)

factor.ice.water.phase(x)

factor.horizontal.averaging(x)

factor.vfm(df)
}
\arguments{
\item{x}{Character or integer.  Vector to convert to factor}

\item{labels}{Character.  Factor labels}

\item{ordered}{Boolean.  Create ordered factor?}

\item{df}{Data.frame.  A data frame containing the variables
\code{feature.qa.lowest.cloud} ... \code{feature.above.surface}
as strings}
}
\value{
With the exception of \code{factor.vfm}, these functions
    return factor version of their character or integer vector
    argument.

\code{factor.vfm} returns a data.frame in which the string
    variables describing the CALIOP VFM have been replaced by
    (ordered, where appropriate) factors
}
\description{
CALIOP VFM factors
}
\section{Functions}{
\itemize{
\item \code{factor.relabel}: Utility function for converting
character or integer vectors to factor

\item \code{factor.feature.type}: Convert feature type from string to factor

\item \code{factor.qa}: Convert feature type from string to factor

\item \code{factor.ice.water.phase}: Convert feature type from string to factor

\item \code{factor.horizontal.averaging}: Convert feature type from string to factor

\item \code{factor.vfm}: Convert CALIOP VFM strings to factors in a data.frame
}}

