% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbase}
\alias{cbase}
\title{CBASE retrievals (2007 and 2008)}
\format{A data frame with \eqn{n} variables:
\describe{
  \item{lon}{longitude; -180 < lon < 180}
  \item{lat}{latitude}
  \item{denom}{denominator for fractions (character):
    \describe{
       \item{all}{number of satellite overpasses}
       \item{cloud}{number of cloudy satellite overpasses}
       \item{rain}{number of raining satellite overpasses}
    }}
}}
\source{
\code{\link{get.cbase.bitsies}}
}
\usage{
cbase
}
\description{
A dataset containing the CBASE cloud base data
}
\keyword{datasets}
