% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbase.tune.R
\name{CBASE_combine}
\alias{CBASE_combine}
\alias{cbase.combine.station}
\alias{cbase.combine.segment}
\alias{cbase.combine}
\title{CBASE combination of local cloud bases into CBASE product}
\usage{
cbase.combine.station(df.cor)

cbase.combine.segment(df.cor)

cbase.combine(df.cor)
}
\arguments{
\item{df.cor}{Data.frame.  Contains (corrected) local cloud bases}
}
\value{
Data frame with combined cloud bases
}
\description{
CBASE combination of local cloud bases into CBASE product
}
\section{Functions}{
\itemize{
\item \code{cbase.combine.station}: Combination around a ground station
overpass

\item \code{cbase.combine.segment}: Combination within a Calipso track
segment

\item \code{cbase.combine}: Actual worker function
}}

