% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbase.tune.R
\name{CBASE_correct}
\alias{CBASE_correct}
\alias{correct.cbase.lm}
\alias{correct.cbase.ident}
\title{CBASE local cloud base correction functions}
\usage{
correct.cbase.lm(df, correction)

correct.cbase.ident(df, correction)
}
\arguments{
\item{df}{Data.frame.  Contains local cloud bases to be corrected}

\item{correction}{List produced by the \code{\link{CBASE_tune}}
functions}
}
\value{
Data frame with corrected local cloud bases
}
\description{
CBASE local cloud base correction functions
}
\section{Functions}{
\itemize{
\item \code{correct.cbase.lm}: Correct using linear model or inherited
classes (including SVM)

\item \code{correct.cbase.ident}: "Identity" (i.e., no) correction
}}

