% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbase.combine.R, R/cbase.gen.R
\name{bases.cbase.combine}
\alias{bases.cbase.combine}
\alias{CBASE_gen}
\alias{bases.cbase}
\alias{bases.cbase.dardar}
\alias{bases.2b.geoprof.lidar}
\title{Combination of local thin-cloud CALIOP bases into CBASE product}
\usage{

  bases.cbase.combine(local.cbase = dbtools::db_spec("cloud-bases-2008-local.sqlite",
  "cloudbase"), ncores = 72)

bases.cbase(path = "/home/jmuelmen/CALIOP/VFM.v4.10/2008",
  pattern = "CAL_LID_L2_VFM-Standard-V4-10.*hdf",
  out.name = "cloud-bases.rds", combination = TRUE,
  cor.svm.fname = "~/models.svm.rds")


  bases.cbase.dardar(path = "/projekt3/climate/DATA/SATELLITE/MULTI_SENSOR/DARDAR/DARDAR_MASK/2007",
  out.name = "cloud-bases.rds")


  bases.2b.geoprof.lidar(path = "/projekt3/climate/DATA/SATELLITE/CLOUDSAT/2B-GEOPROF-LIDAR/2008")
}
\arguments{
\item{local.cbase}{Database containing the local cloud bases}

\item{path}{Character.  Apply algorithm to all HDF files in this path}

\item{out.name}{Character.  Output file name}
}
\value{
A data frame containing information from both

A data frame containing cloud bases
}
\description{
Combination of local thin-cloud CALIOP bases into CBASE product

CBASE generator functions
}
\section{Functions}{
\itemize{
\item \code{bases.cbase.combine}: Combine local

\item \code{bases.cbase}: Bases by non-attenuated CALIOP in thin
clouds (using CALIOP VFM)

\item \code{bases.cbase.dardar}: Bases by non-attenuated CALIOP in thin clouds (using DARDAR mask)

\item \code{bases.2b.geoprof.lidar}: Bases by 2B-GEOPROF-LIDAR CloudSat/CALIOP combination
}}

