% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbase.tune.R
\name{CBASE_tune}
\alias{CBASE_tune}
\alias{tune.cbase.lm}
\alias{tune.cbase.svm}
\title{CBASE tuning functions}
\usage{
tune.cbase.lm(df, thresh = NULL)

tune.cbase.svm(df)
}
\arguments{
\item{df}{Data.frame.  Contains tuning data set}

\item{thresh}{Numeric.  If non-NULL, exclude retrieved base heights
below \code{thresh} from the fit.}
}
\value{
A list with two components, a list of tuned models for each
    data category and a data.frame containing the index into said list of models
}
\description{
CBASE tuning functions
}
\section{Functions}{
\itemize{
\item \code{tune.cbase.lm}: Tuning using linear models

\item \code{tune.cbase.svm}: Tuning using SVM
}}

