% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbasetools.R
\docType{package}
\name{cbasetools-package}
\alias{cbasetools}
\alias{cbasetools-package}
\title{Retrieve and evaluate cloud bases from various satellite-borne
instruments}
\description{
Retrieve and evaluate cloud bases from various satellite-borne
instruments
}

