% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine.cbase.metar}
\alias{combine.cbase.metar}
\title{Combine METAR with cloud base retrievals}
\usage{
combine.cbase.metar(eval = get.metar.2008(),
  retrieval = dbtools::db_spec("cloud-bases-2008.sqlite", "cloudbase"),
  resolution, resolution_names, ncores = 72)
}
\description{
Combine METAR with cloud base retrievals
}

