% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine.cbase}
\alias{combine.cbase}
\title{Combination of retrievals for comparison}
\usage{
combine.cbase()
}
\arguments{
\item{df.cbase}{Data.frame.  Data frame to be evaluated.}

\item{df.eval}{Data.frame.  Data frame to evaluate against.}
}
\value{
A data frame containing information from both
}
\description{
Combine METARs with C-BASE retrievals
}

