% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert.date}
\alias{convert.date}
\title{Convert standard time to format to Y/m/d date}
\usage{
convert.date(time)
}
\arguments{
\item{time}{Vector of POSIXct values}
}
\value{
Vector of character in Y/m/d format
}
\description{
Convert standard time to format to Y/m/d date
}
\examples{

}

