% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{df.to.db}
\alias{df.to.db}
\title{Save a data.frame to an SQL database}
\usage{
df.to.db(dd, fname, name, indexes = list("lon", "lat"))
}
\arguments{
\item{dd}{data.frame.}

\item{fname}{character.  Output file name.}

\item{name}{character.  Table name.}

\item{indexes}{list.  List of variables to index.}
}
\value{
\code{dplyr::tbl} version of the database table
}
\description{
Save a data.frame to an SQL database
}
