% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval.efficiency}
\alias{eval.efficiency}
\title{Evaluation function (retrieval efficiency)}
\usage{
eval.efficiency()
}
\arguments{
\item{path}{Character.  Apply algorithm to all HDF files in this path}
}
\value{
A data frame containing cloud bases
}
\description{
Evaluation function (retrieval efficiency)
}

