% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval.metar}
\alias{eval.metar}
\title{Evaluation function (METAR comparison)}
\usage{
eval.metar()
}
\arguments{
\item{path}{Character.  Apply algorithm to all HDF files in this path}
}
\value{
A data frame containing cloud bases
}
\description{
Evaluation function (METAR comparison)
}

