% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{factor.relabel}
\alias{factor.relabel}
\title{Utility function for converting character or integer vectors to
    factor}
\usage{
factor.relabel(x, labels, ordered = FALSE)
}
\arguments{
\item{x}{Character or integer.  Vector to convert to factor}

\item{labels}{Character.  Factor labels}

\item{ordered}{Boolean.  Create ordered factor?}
}
\description{
Utility function for converting character or integer vectors to
    factor
}

