% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar.gen.R
\name{get.metars.from.archive}
\alias{get.metars.from.archive}
\title{Retrieve METARS by date and ICAO code}
\usage{
get.metars.from.archive(stat.dates)
}
\arguments{
\item{stat.dates}{Data frame.  Output of \link{add.metar.episodes}.}
}
\value{
Vector of distances (in km) between points
}
\description{
Remember to run the data frame through \code{mutate(date =
convert.date(time))} first
}
\examples{

}

