% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metar.to.cloud.heights}
\alias{metar.to.cloud.heights}
\title{Parse METAR for cloud heights}
\usage{
metar.to.cloud.heights(metar)
}
\arguments{
\item{metar}{Character.  Standard-format METAR message}
}
\value{
Data frame containing up to three cloud base heights (in
    meters) and the corresponding coverage fractions (FEW|SCT|BKN|OVC)
}
\description{
(up to three layers)
}

