% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R, R/evalutils.R
\name{regression_plot}
\alias{regression_plot}
\title{make a ggplot the way we like them, based on a grouped data frame}
\usage{
regression_plot(df, title, xlab = "50 km running mean CALIOP base (km)")

regression_plot(df, title, xlab = "50 km running mean CALIOP base (km)")
}
\value{

}
\description{
make a ggplot the way we like them, based on a grouped data frame

Make a ggplot the way we like them, based on a grouped data frame
}
\examples{

}

