% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalutils.R
\name{regression_table}
\alias{regression_table}
\title{Make an xtable from statistics of a grouped data frame}
\usage{
regression_table(df.stats)
}
\arguments{
\item{df.stats}{List or data.frame.  Statistics summary from
\code{statistify()} or list output from \code{regression_plot}
(in which case the \code{stats} element is used)}
}
\value{
An \code{xtable}
}
\description{
Make an xtable from statistics of a grouped data frame
}
\examples{

}
