% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{resolution.names}
\alias{resolution.names}
\title{Get resolution names from function or function list}
\usage{
resolution.names(resolution)
}
\arguments{
\item{resolution}{Function or list of functions}
}
\description{
Get resolution names from function or function list
}

