% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{running.gaussian}
\alias{running.gaussian}
\title{Apply gaussian filter to vector}
\usage{
running.gaussian(x, half.len, sd)
}
\value{
Data frame
}
\description{
Apply gaussian filter to vector
}
\examples{

}

