% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalutils.R
\name{statistify.df}
\alias{statistify.df}
\title{Add summary statistics (linear regression, correlation, RMS error)
to a data frame or grouped data frame}
\usage{
statistify.df(df)
}
\value{

}
\description{
Add summary statistics (linear regression, correlation, RMS error)
to a data frame or grouped data frame
}
\examples{

}
