% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R, R/evalutils.R
\name{statistify}
\alias{statistify}
\title{add summary statistics (linear regression, correlation, RMS error) to a data frame or grouped data frame}
\usage{
statistify(df)

statistify(df)
}
\value{

}
\description{
add summary statistics (linear regression, correlation, RMS error) to a data frame or grouped data frame

Add summary statistics (linear regression, correlation, RMS error)
to a data frame or grouped data frame
}
\examples{

}

