% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim}
\alias{trim}
\title{Trim whitespace from string}
\usage{
trim(x)
}
\arguments{
\item{x}{Character}
}
\value{
Character without leading or trailing whitespace
}
\description{
Trim whitespace from string
}
