% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visutils.R
\name{vis.cdf}
\alias{vis.cdf}
\title{Visualization of geographical distribution of cloud-base
uncertainties (CDF)}
\usage{
vis.cdf(df)
}
\arguments{
\item{df}{data.frame.  Result of \code{bases.cbase()}}
}
\description{
Visualization of geographical distribution of cloud-base
uncertainties (CDF)
}
