% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visutils.R
\name{vis.diff}
\alias{vis.diff}
\title{Visualization of geographical distribution of cloud base heights
(day-night differences)}
\usage{
vis.diff(df, base_size)
}
\arguments{
\item{df}{data.frame.  Result of \code{bases.cbase()}}
}
\description{
Visualization of geographical distribution of cloud base heights
(day-night differences)
}
