% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R, R/visutils.R
\name{vis.geo}
\alias{vis.geo}
\alias{vis.geo}
\title{Visualization of geographical distribution of cloud bases}
\usage{
vis.geo(df)

vis.geo(df)
}
\arguments{
\item{df}{data.frame.  Result of \code{bases.cbase()}}

\item{df}{data.frame.  Result of \code{bases.cbase()}}
}
\description{
Visualization of geographical distribution of cloud bases

Visualization of geographical distribution of cloud bases
}
