% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visutils.R
\name{vis.mean}
\alias{vis.mean}
\title{Visualization of geographical distribution of cloud base heights}
\usage{
vis.mean(df, base_size)
}
\arguments{
\item{df}{data.frame.  Result of \code{bases.cbase()}}
}
\description{
Visualization of geographical distribution of cloud base heights
}
