% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c-base.gen.R
\name{C-BASE_gen}
\alias{C-BASE_gen}
\alias{bases.2b.geoprof.lidar}
\alias{bases.cbase}
\title{C-BASE generator functions}
\usage{

  bases.cbase(path = "/projekt3/climate/DATA/SATELLITE/MULTI_SENSOR/DARDAR/DARDAR_MASK/2007")


  bases.2b.geoprof.lidar(path = "/projekt3/climate/DATA/SATELLITE/CLOUDSAT/2B-GEOPROF-LIDAR/2008")
}
\arguments{
\item{path}{Character.  Apply algorithm to all HDF files in this path}
}
\value{
A data frame containing cloud bases
}
\description{
C-BASE generator functions
}
\section{Functions}{
\itemize{
\item \code{bases.cbase}: Bases by non-attenuated CALIOP in thin clouds

\item \code{bases.2b.geoprof.lidar}: Bases by 2B-GEOPROF-LIDAR CloudSat/CALIOP combination
}}

