% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine.cbase}
\alias{combine.cbase}
\title{Combination of retrievals for comparison}
\usage{
combine.cbase(df.eval, df.ret, max.dist)
}
\arguments{
\item{df.eval}{Data.frame.  Data frame to evaluate against.}

\item{df.ret}{Data.frame.  Data frame of retrievals to be evaluated.}

\item{max.dist}{Numeric.  Maximum distance between retrieval and evaluation points}
}
\value{
A data frame containing information from both
}
\description{
Combine evaluation dataset with retrievals to be evaluated
}

