% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalutils.R
\name{regression_plot}
\alias{regression_plot}
\title{Make a ggplot the way we like them, based on a grouped data frame}
\usage{
regression_plot(df, title, xlab = "CALIOP base retrieval (km)")
}
\value{

}
\description{
Make a ggplot the way we like them, based on a grouped data frame
}
\examples{

}

