% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{resolution}
\alias{resolution}
\alias{resolution.all}
\alias{resolution.min_cbh}
\alias{resolution.min_cbh.qual}
\alias{resolution.min_dist}
\alias{resolution.quantile_cbh.qual}
\title{Methods to resolve multiple retrievals at the same location}
\usage{
resolution.all(df)

resolution.min_dist(df)

resolution.min_cbh(df)

resolution.min_cbh.qual(df)

resolution.quantile_cbh.qual(prob)
}
\arguments{
\item{df}{Data frame containing multiple retrievals}

\item{prob}{Scalar numeric.  Quantile of cloud base altitude}
}
\value{
A \code{data.frame} containing the resolution of the
    multiple retrievals (single retrieval)
}
\description{
Methods to resolve multiple retrievals at the same location
}
\section{Functions}{
\itemize{
\item \code{resolution.all}: All retrieval within 100 km

\item \code{resolution.min_dist}: Retrieval at minimum distance

\item \code{resolution.min_cbh}: Retrieval at minimum cloud base height

\item \code{resolution.min_cbh.qual}: Retrieval at minimum cloud base height subject to quality criteria

\item \code{resolution.quantile_cbh.qual}: Retrieval at quantile of cloud base height
}}

