% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visutils.R
\name{vis.quantiles}
\alias{vis.quantiles}
\title{Visualization of geographical distribution of cloud-base
uncertainties (quantiles)}
\usage{
vis.quantiles(df)
}
\arguments{
\item{df}{data.frame.  Result of \code{bases.cbase()}}
}
\description{
Visualization of geographical distribution of cloud-base
uncertainties (quantiles)
}
